+++ package.json
@@ -12,6 +12,7 @@
     "express": "^4.17.1",
     "install": "^0.13.0",
     "mysql": "^2.18.1",
+    "node-sass": "^4.14.1",
     "nodemon": "^2.0.3",
     "path": "^0.12.7",
     "pg": "^8.1.0",
