+++ src/assets/css/_print.scss
@@ -1,0 +1,25 @@
+.packing-list-page {
+  @media print {
+    span, p {
+      font-size: .75rem;
+    }
+
+    div {
+      page-break-inside: avoid;
+    }
+
+    .check-box {
+      padding: 0;
+    }
+
+    .print-button {
+      display: none;
+    }
+  }
+}
+@page {
+  margin: 0;
+  body  {
+    margin: 1.5cm;
+  }
+}
