+++ src/assets/css/main.scss
@@ -1,3 +1,5 @@
+@import './_print.scss';
+
 body {
   background-color: #f1f1ef;
 }
